% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercnvPlus.R
\name{calcCNV}
\alias{calcCNV}
\title{Calculate CNV scores given a matrix of RNASeq counts. Output a matrix of final values.}
\usage{
calcCNV(
  data,
  gene_pos,
  cutoff,
  reference_obs,
  window_size,
  out_path,
  contig_tail,
  noise_filter,
  vis_bounds
)
}
\arguments{
\item{data:}{expression matrix (genes X samples),
assumed to be log2(TPM+1) .}

\item{gene_order:}{ordering of the genes (data's rows)
according to their genomic location
To include all genes use 0.}

\item{cutoff:}{cut-off for the average expression of genes to be
used for CNV inference.}

\item{reference_obs:}{Column names of the subset of samples (data's columns)
that should be used as references.
If not given, the average of all samples will
be the reference.}

\item{window_size:}{length of the window for the moving average
(smoothing). Should be an odd integer.}

\item{out_path:}{the path to what to save the pdf as. The raw data is
also written to this path but with the extension .txt .}

\item{contig_tail:}{length of the tail removed from the ends of contigs.}

\item{noise_filter:}{the minimum difference a value can be from the
average reference in order for it not to be
removed as noise.}

\item{vis_bounds:}{Used as upper and lower bounds for values in the visualization.
Should be given in the form of '-1,1' (lower bound, upper bound).}
}
\value{
Returns an 'infercnv' object including:
    1. CNV matrix before visualization.
    2. CNV matrix after denoise and outlier removal for visualization.
    3. Chromosome order.
    4. Names of cells in reference groups.
}
\description{
Calculate CNV scores given a matrix of RNASeq counts. Output a matrix of final values.
}
