% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractCells}
\alias{extractCells}
\title{Subtract cells from specific branches}
\usage{
extractCells(data, subtrees = NULL, lab_to_rm = "ref")
}
\arguments{
\item{data}{an 'infercnv' object as produced by inferCNV.}

\item{subtrees}{a numeric scalar or vector specify those subtrees to extract.}

\item{lab_to_rm}{a character scalar with the label of cells to exclude.}
}
\value{
Return an 'infercnv' object with target cells
}
\description{
Args:
}
