% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gtf_to_position}
\alias{gtf_to_position}
\title{Generate a genomic positions file from a GTF file}
\usage{
gtf_to_position(gtf, out_file = "genomic_positions.txt", out_path = "./")
}
\arguments{
\item{gtf}{path to input gtf file.}

\item{out_file}{filename to save genomic positions.}

\item{out_path}{output directory.}
}
\value{
Return genomic positions.
}
\description{
Args:
}
