% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pruneCutree}
\alias{pruneCutree}
\title{Prune cutree to dendlist

Args:}
\usage{
pruneCutree(data, k, out_path = "output_dir")
}
\arguments{
\item{data}{an 'infercnv' object as produced by inferCNV.}

\item{k}{an integer scalar or vector with the desired number of groups.}

\item{out_path}{output directory.}
}
\value{
Return an 'infercnv' object with dendlist.
}
\description{
Prune cutree to dendlist

Args:
}
