% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{umi_to_log2tpm}
\alias{umi_to_log2tpm}
\title{Transform UMI to log2(TPM+1)}
\usage{
umi_to_log2tpm(data)
}
\arguments{
\item{data}{UMI counts matrix (genes X cells).}
}
\value{
Return log2-transformed TPM matrix.
}
\description{
Args:
}
