% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{visualCNV}
\alias{visualCNV}
\title{Clustering and plotting cells based on cnv score matrix using ComplexHeatmap package.

Args:}
\usage{
visualCNV(
  data,
  assay = NULL,
  ref_lab = "ref",
  obs_lab = "obs",
  distance_method = "euclidean",
  clustering_method = "ward.D2",
  cutree_k = NULL,
  colors = colorRampPalette(rev(brewer.pal(n = 11, name = "RdBu")))(11),
  border = TRUE,
  out_file = "plot_cnv.png",
  out_path = "output_dir"
)
}
\arguments{
\item{data}{an 'infercnv' object as produced by inferCNV.}

\item{assay}{assay which assigned to use for plotting.}

\item{ref_lab}{label for reference cells.}

\item{obs_lab}{label for observations.}

\item{clustering_method}{clustering method used. Accepts the same values as stats::hclust.}

\item{cutree_k}{an integer scalar or vector with the desired number of groups.}

\item{colors}{vector of colors used in heatmap.}

\item{border}{whether draw border. The value can be logical or a string of color.}

\item{out_file}{filename to save plot.}

\item{out_path}{output directory.}

\item{dist_method}{distance measure used in clustering cells, possible
values are 'correlation' for Pearson correlation 
and all the distances supported by stats::dist.}
}
\value{
An 'infercnv' object
}
\description{
Clustering and plotting cells based on cnv score matrix using ComplexHeatmap package.

Args:
}
