% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importSrat}
\alias{importSrat}
\title{Import a seurat object and convert it to matrix for inferCNV.}
\usage{
importSrat(obj, slot = "counts", assay = "RNA")
}
\arguments{
\item{slot}{slot to use}

\item{assay}{which assays to use.}

\item{data}{a Seurat object.}
}
\value{
Return a TPM log2-transformed matrix.
}
\description{
Args:
}
